Don't use a .templates.h file if we unconditionally include it anyway.
